r176: more robust to cross-species uses
Now pangene by default elevates the score of a protein if it overlaps with
a better-aligned protein not considered in the vertex set. This appears to
improve the human-primate-mouse graph. On human data, this change has little
effect. All detected bubbles remain the same.

At one point, I was thinking to cluster proteins. A problem is that orthologous
genes can't be transferred over species. The current algorithm might be better.
We will see on more datasets.